+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2026-07-10]
+## [1.0.0 - 2026-07-14]
 ### Added
 - First version of the project
 - Spring Application
@@ -267,6 +267,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Delete Form Field Block API endpoint
 - Form field properties
 - Normalize paths in constructor parameters
+- Condition in form field config
 
 
 ### Changed